requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
cloudscraper==1.2.40
fake-useragent==0.1.11
idna==2.9
lxml==4.5.1
mysql-connector-python==8.0.20
protobuf==3.12.2
pyparsing==2.4.7
pyperclip==1.8.0
PySocks==1.7.1
requests==2.24.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
